server/requirements.txt
fastapi
uvicorn
motor
requests
GitPython
